+++ device-management-application/CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - ET-Top: Add 'w' interactive command to toggle display of watts in horizonal bar form.
 - ET-Top: Add labels to the display of watts in horizonal bar form.
+- ET-Top: Add MISC power at a constant 5.75 watts under the ETSOC power category.
 ### Changed
 ### Deprecated
 ### Removed